# swagger-auth.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/roles/{id}/privilegios/{privilegioId}": {
            "delete": {
                "operationId": "RolesController_removePrivilegio",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "privilegioId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Privilegio removido del rol"
                    }
                },
                "summary": "Remover un privilegio de un rol",
                "tags": [
                    "Roles"
                ]
            }
        },
        "/privilegios": {
            "post": {
                "operationId": "PrivilegiosController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePrivilegioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Privilegio creado exitosamente"
                    },
                    "409": {
                        "description": "El privilegio ya existe"
                    }
                },
                "summary": "Crear un nuevo privilegio",
                "tags": [
                    "Privilegios"
                ]
            },
            "get": {
                "operationId": "PrivilegiosController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Lista de privilegios"
                    }
                },
                "summary": "Obtener todos los privilegios",
                "tags": [
                    "Privilegios"
                ]
            }
        },
        "/privilegios/{id}": {
            "get": {
                "operationId": "PrivilegiosController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Privilegio encontrado"
                    },
                    "404": {
                        "description": "Privilegio no encontrado"
                    }
                },
                "summary": "Obtener un privilegio por ID",
                "tags": [
                    "Privilegios"
                ]
            },
            "patch": {
                "operationId": "PrivilegiosController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePrivilegioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Privilegio actualizado"
                    },
                    "404": {
                        "description": "Privilegio no encontrado"
                    }
                },
                "summary": "Actualizar un privilegio",
                "tags": [
                    "Privilegios"
                ]
            },
            "delete": {
                "operationId": "PrivilegiosController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Privilegio eliminado"
                    },
                    "404": {
                        "description": "Privilegio no encontrado"
                    }
                },
                "summary": "Eliminar un privilegio",
                "tags": [
                    "Privilegios"
                ]
            }
        },
        "/alumnos": {
            "post": {
                "operationId": "AlumnosController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAlumnoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alumno creado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "409": {
                        "description": "El usuario ya tiene perfil de alumno o código duplicado"
                    }
                },
                "summary": "Crear un nuevo alumno",
                "tags": [
                    "Alumnos"
                ]
            },
            "get": {
                "operationId": "AlumnosController_findAll",
                "parameters": [
                    {
                        "name": "escuela",
                        "required": false,
                        "in": "query",
                        "description": "Filtrar por ID de escuela",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de alumnos"
                    }
                },
                "summary": "Obtener todos los alumnos",
                "tags": [
                    "Alumnos"
                ]
            }
        },
        "/alumnos/{id}": {
            "get": {
                "operationId": "AlumnosController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumno encontrado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    }
                },
                "summary": "Obtener un alumno por ID",
                "tags": [
                    "Alumnos"
                ]
            },
            "patch": {
                "operationId": "AlumnosController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAlumnoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alumno actualizado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    }
                },
                "summary": "Actualizar un alumno",
                "tags": [
                    "Alumnos"
                ]
            },
            "delete": {
                "operationId": "AlumnosController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Alumno eliminado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    }
                },
                "summary": "Eliminar un alumno",
                "tags": [
                    "Alumnos"
                ]
            }
        },
        "/alumnos/usuario/{usuarioId}": {
            "get": {
                "operationId": "AlumnosController_findByUsuario",
                "parameters": [
                    {
                        "name": "usuarioId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumno encontrado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    }
                },
                "summary": "Obtener alumno por usuario ID",
                "tags": [
                    "Alumnos"
                ]
            }
        },
...
    },
    "info": {
        "title": "PPP Auth Service API",
        "description": "Autenticación y Autorización - Usuarios, Roles, Privilegios",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "CreatePrivilegioDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "description": "Nombre del privilegio",
                        "example": "crear_usuario",
                        "maxLength": 100
                    }
                },
                "required": [
                    "nombre"
                ]
            },
            "UpdatePrivilegioDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "description": "Nombre del privilegio",
                        "example": "crear_usuario",
                        "maxLength": 100
                    }
                }
            },
            "CreateAlumnoDto": {
                "type": "object",
                "properties": {
                    "usuarioId": {
                        "type": "string",
                        "description": "ID del usuario asociado",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "idEscuela": {
                        "type": "string",
                        "description": "ID de la escuela profesional",
                        "example": "123e4567-e89b-12d3-a456-426614174001"
                    },
                    "codigo": {
                        "type": "string",
                        "description": "Código del alumno",
                        "example": "2020-123456",
                        "maxLength": 30
                    },
                    "ciclo": {
                        "type": "string",
                        "description": "Ciclo actual del alumno",
                        "example": "VIII",
                        "maxLength": 20
                    },
                    "año": {
                        "type": "string",
                        "description": "Año de ingreso",
                        "example": "2020",
                        "maxLength": 4
                    }
                },
                "required": [
                    "usuarioId",
                    "idEscuela",
                    "codigo",
                    "ciclo",
                    "año"
                ]
            },
            "UpdateAlumnoDto": {
                "type": "object",
                "properties": {
                    "idEscuela": {
                        "type": "string",
                        "description": "ID de la escuela profesional",
                        "example": "123e4567-e89b-12d3-a456-426614174001"
                    },
                    "codigo": {
                        "type": "string",
                        "description": "Código del alumno",
                        "example": "2020-123456",
                        "maxLength": 30
                    },
                    "ciclo": {
                        "type": "string",
                        "description": "Ciclo actual del alumno",
                        "example": "VIII",
                        "maxLength": 20
                    },
                    "año": {
                        "type": "string",
                        "description": "Año de ingreso",
                        "example": "2020",
                        "maxLength": 4
                    }
                }
            },
...
        }
    }
}